+++ src/routes/(app)/+layout.svelte
@@ -57,9 +57,7 @@
   bovedin={settings.bovedin}
 />
 <div
-  class="selection:bg-fuchsia-300 selection:text-fuchsia-900"
-  in:fade={{ duration: 300, delay: 300 }}
-  out:fade={{ duration: 100, delay: 0 }}
+  class="selection:bg-fuchsia-300 selection:text-fuchsia-900 h-full"
 >
   <slot />
 </div>