+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix result Youtube default iamge
+- Fix next event on home
+
 ## [0.0.11] - 2024-01-07
 
 ### Fixed